Introduce a new handshake message between the Probe and the Collector. This allows to differentiate between Probe with no data to report and probe with proben configuration. The Handshake message sends a ProbeDetails xml record. The Handshake is done the first time (in a process) Gratia.Initiliaze is called. The Handshake can also be called explicitly:    Gratia.Handshake() This is in particular recommended for Probe that are implemented as a daemon, which would call Handshake regularly (especially if there is no real data to upload).  The call to Handshake will also re-establish the connection if it was previously broken or it has timed-out.
The ProbeDetails contains information about the Probe software releases
and about the Service (being probe) software release.

Probe should call:
  Gratia.RegisterReporterLibrary(name,version)
to register the libraries the Probe itself is using.
  Gratia.RegisterReporter(name,version)
to register the version of the Probe itself
  Gratia.RegisterService(name,version)
to register the version of the service (Condor,
PBS,LSF,DCache,SunGrid Engine, etc.) being 'probed'/

Those function must be called before the call to Gratia.Initialize.

ProbeDetails xml files are __not__ cached if the the upload fails.