Facilities for configuration to use Tomcat Native Libraries (org.apache.coyote.http11.Http11AprProtocol instead of org.apache.coyote.http11.Http11BaseProtocol). Simply configure
tomcat_native_lib => '/data/apr/lib'

(or whatever).

Note that the tomcat-native libraries as built from the tar file in
tomcat/bin must be made aware at build-time of the location of the
Apache Portable Runtime libraries and includes. The easiest way of doing
this is to put them in the same place.

See gratia-vm02:/data/apr for illustration.