Introduce a new property gratia.database.useJobUsageSiteName. Like gratia.database.wantSummaryTable it is read from the property file and stored in the Database.
useJobUsageSiteName indicates that the 'ReportedSiteName' should be considered part of the data (for md5 calculation) while it is usually only an indicative field.

wantSummaryTable is used to avoid attempting to clean the summary table when it does not exist.

Both value are now also cached in a static variable of DatabaseMaintenance.