Implement Probe to Collector authentication and Origin Tracking.
There are 5 levels of security checks and origin tracking which can be configured via the services-configuration.properties file:

# service.security.level -> security level/provider gratia uses
#   value: 0 -> no security [Current default]
#   value: 1 -> Track connections without SSL, check connection validity
#   value: 2 -> Check only SSL certificate (both cert validity and validity within Collector)
#   value: 3 -> Same as 2 + Track connection with SSL and check connection validity
#   value: 4 -> Require SSL connection + same checks as 2
#   value: 5 -> Same as 4 + Track connection with SSL and check connection validity

The information related to the certificate is stored in the Certificate table which includes the PEM of the certificate. 
The information related to the 'client' is stored in the Connection table which contains information about the sender and its certificate.
The 'Origin' table contains the information about a specific upload, i.e the time of the connection and a pointer to the connection information.
The JobUsageRecord_Origin table contains the link between a specific record and its origins, including the 'hop' number of the origin.

AttachableXmlElement and AttachableCollection formalize the concept of objects that needs to be attached to an incoming record but that we cache for longer than the duration of an hibernation session/transaction in order avoid having to look them up from the database for each record (that need them).   Currently 'Software' and 'Connection' are using this mechanism.  This needs to be check/reviewed when enabling more than one thread.

Related Updates:

New classes: Origin,Connection,Certificate,OriginLoader,AttachableXmlElement,AttachableCollection

The ListenerThread has been extended to properly handle incoming 'origin' information.
The Collector services now has a 'name' and implements the connection checking (CheckConnection).

The Record.asXML now takes a parameter to select where the XML is for storing in the DB or for md5 calculation.  This avoids the manipulation needed to 'skip' some information in the xml needed for md5 calculation.

Factored out a log of functionality that was duplicated in ProbeDetails, JobUsageRecord and MetricRecord (See the Record and RecordLoader classes for details)

Fixed a problem when writing the XML datum for a Duration lasting more than 1 month.

Fixed some issues with the replication to file.

Severals fixes/extensions to file layout and log messages.

Upcoming (and necessary before the next release):

add proper indices, add proper table clean, extend origin table to ProbeDetails and MetricRecord