Config:get_MaxPendingFiles: now return an int to allow proper comparisons. BundleSize: now limited to the min of the requested size and MaxPendingFiles Class Response: fix initializer in case the given code is not -1 Add HasMoreOutstandingRecord keeping track of whether SearchOutstandingRecord exited 'early' Add OutstandingRecordCount keeping track of the number of files in the gratiafiles directory
Reprocess now attempts to send all the outstanding record before proceeding.

Reprocess now returned a simplier/smaller output string:
    OK - Reprocessing 0 record(s) uploaded, 1 bundled, 0 failed

Fix several issue essentially disabling the limitation of the number of files in 
gratiafiles.

Even when there are too many pending files, in Send, attempt to send the record
to the collector; only in the case where the upload fails do we complain that
there are too many files.

Send now attempts to Reprocess not only just after recovering from a connection
error but also when the last SearchOutstandindRecords did not use up all the
pending records.