Import revision 3783 from the trunk:
Optimize the deletion of the RawXML rows by insuring the delete  does NOT use the clause 'dbid in (select ...)'
when using mysql.  This is currenlty horrendously slow with large tables (most likely 
incurring a scan of the large table for some reason).  Instead we use the delete with
multi-table syntax.