runPurge: add -b option to do both the build and upload of a war file
New class QueueManager::File to encapsulate the path, origin and number 
of records of one of the input file (and the way to calculate the last
when needed).

Add a string Status to QueueManager
Refactor a few routines in QueueManager (leveraging the File class).
Rename resetStatus to refreshStatus

Add way to force the refreshing of the QueueManager status from
the system administration page.

Use the number status informatin in the queue monitoring pages.

In RecordProcessor fix SaveQuarantine to also properly update the
queue status (this is the part needing the new File class).