The QSizeMonitor now pause the housekeeping service if the number of record in one of the queue reach the value of 'max.housekeeping.nrecords' in the configuration file.  The housekeeping service is restarted if the number of record in each queues goes back under 'min.housekeeping.nrecords'
Add test of the queue size monitor handling of the servlet and the housekeeping.

Introduce a time waster portion in the housekeeping which is activiate only 
if the housekeeping has nothing to do AND the ssl port and the regular port
are the same (true only in the test environment).

Add a way to force the running of the QSizeMonitor (to speed up the test).

Simplify runPurge.sh by adding more functions