Probes to record the status of the BDII.
The gip-status probes records subclusters, CEs, and SEs found in the BDII.
These probes have existed for awhile at Nebraska and used to record the status
found in the OSG BDII; committed here are the cleaned-up version of the probes.

The CE and SE status are used for OSG-Metrics internal graphs, but the subcluster
information is recorded daily and used by the OSG's monthly WLCG accounting to
calculate normalization constants.