Complete rework of the way we gather and display the statistics and performance information for the Collector
The Status and MonitorStatus administration pages have been updated to take advantage of the
new tables and are now fast even under heavy load.

4 tables now participates: TableStatistics, TableStatisticsSnapshots, TableStatisticsHourly, TableStatisticsDaily

TableStatistics contains for each record type (JobUsage, Metric, etc.) the number of records that
are currently in the database and the number of records that have been processed (even if they have
been deleted since).   For each, the table not only the number of good record but also the number
of duplicates and the number of 'errorneous' records ('Qualifier').
The main new feature in this table is the addition of the 'lifetime' number in addition to the current numbers.

TableStatisticsSnapshots contains a copy of the content of TableStatistics taken every five minutes (customizable)
and kept for a day (default)

TableStatisticsHourly contains a summary of the snapshots taken during a given hour.   For each entry
(current/lifetime,RecordType,Qualifier), the table records the exact start and end time it covers,
the minimum, maximum and average number of records during the period.  The default is to keep those
hourly summary for one year.

TableStatisticsDaily contains a summary of the snapshots taken during a given day.   For each entry
(current/lifetime,RecordType,Qualifier), the table records the exact start and end time it covers,
the minimum, maximum and average number of records during the period.  The default is to keep those
daily summary indefinitely. 


A new class, TableStatisticsManager, is introduced to manage the lifetime of the rows in the new
Statistics tables and to do/schedule the copying and summarizing.  2 new mysql function are also
introduced for this purpose: table_statictics_hourly_summary and table_statictics_daily_summary

The Backlog class and table are disabled (and probably obsolete).

New configuration items:

# frequency of the snapshots, 0 disable the history recording.
# In number of minutes between snapshots
#
tableStatistics.snapshots.wait = 5
#
# How long to keep the individual snap shots of the table statistics
#
service.lifetime.TableStatisticsSnapshots = 1 day
#
# How long to keep the hourly summary of the table statistics
#
service.lifetime.TableStatisticsHourly = 1 year
#
# How long to keep the daily summary of the table statistics
#
service.lifetime.TableStatisticsDaily = UNLIMITED
#
# input queue control - whether or not to monitor queue sizes.
#